requirements.txt
accelerate==0.29.1
albumentations==1.4.3
aniso8601==9.0.1
ansi2html==1.9.1
antlr4-python3-runtime==4.9.3
anyio
argon2-cffi
argon2-cffi-bindings
arrow
asttokens
async-lru
attrs
autovizwidget==0.21.0
awscli==1.32.78
Babel
beautifulsoup4
bleach
blinker==1.7.0
bokeh
boto3==1.34.78
botocore==1.34.78
Brotli
cached-property
captum
certifi
cffi
charset-normalizer
click==8.1.7
cloudpickle==2.2.1
colorama
colorlog==6.8.2
comm
contextlib2==21.6.0
contourpy
cryptography
cycler
debugpy
decorator
defusedxml
dill==0.3.8
docker==6.1.3
docutils
dparse==0.6.3
einops==0.7.0
entrypoints
environment-kernels==1.2.0
exceptiongroup
executing
fastjsonschema
filelock
Flask==3.0.2
Flask-RESTful==0.3.10
fonttools
fqdn
fsspec==2024.2.0
gmpy2
google-pasta==0.2.0
gssapi==1.8.3
gym==0.26.2
gym-notices==0.0.8
h11
h2
hdijupyterutils==0.21.0
hpack==4.0.0
httpcore
httpx
huggingface-hub==0.22.2
hyperframe
idna
imageio
importlib-metadata==6.11.0
importlib-resources
ipykernel
ipython
ipywidgets
isoduration
itsdangerous==2.1.2
jedi
Jinja2
jmespath
joblib==1.3.2
json5
jsonpointer
jsonschema
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events
jupyter-lsp
jupyter_client
jupyter_core
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
krb5==0.5.1
lazy_loader==0.4
llvmlite==0.42.0
loguru==0.7.2
MarkupSafe
matplotlib
matplotlib-inline
mistune
mpi4py==3.1.5
mpmath
multiprocess==0.70.16
munkres==1.1.4
nbclient
nbconvert
nbformat
nest_asyncio
networkx
notebook
notebook_shim
numba==0.59.0
numpy
nvgpu==0.10.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-ml-py==12.535.133
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
omegaconf==2.3.0
onnx
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
orjson==3.10.0
overrides
packaging==21.3
pandas==1.5.3
pandocfilters
parso
pathos==0.3.2
patsy
pexpect
pickleshare
pillow
pkgutil_resolve_name
platformdirs
plotly==5.19.0
pox==0.3.4
ppft==1.7.6.8
prometheus_client
prompt-toolkit
protobuf==4.25.3
psutil
ptyprocess
pure-eval
py4j==0.10.9.5
pyarrow==15.0.0
pyasn1
pybind11
pybind11-global
pycparser
pydantic
pydantic_core
pyfunctional==1.4.3
pygame==2.5.2
Pygments
pynvml==11.5.0
pyparsing
PySocks
pyspark==3.3.0
pyspnego==0.10.2
python-dateutil
python-json-logger
pytz
PyYAML
pyzmq
qtconsole==5.5.1
QtPy==2.4.1
referencing
requests
requests-kerberos==0.14.0
rfc3339-validator
rfc3986-validator
rpds-py
rsa
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
s3fs==0.4.2
s3transfer
safetensors==0.4.2
sagemaker==2.214.3
sagemaker_pyspark==1.4.5
schema==0.7.5
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy
seaborn
Send2Trash
shap==0.40.0
six
slicer==0.0.7
smclarify==0.5
smdebug-rulesconfig==1.0.1
sniffio
soupsieve
sparkmagic==0.21.0
stack-data
statsmodels
sympy
tabulate==0.9.0
tblib==2.0.0
tenacity==8.2.3
termcolor==2.4.0
terminado
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.16
tinycss2
tomli
torch==2.1.0
torch-model-archiver
torch-workflow-archiver
torchaudio==2.1.0
torchdata
torchserve
torchtext==0.16.0
torchvision==0.16.0
tornado
tqdm
traitlets
triton==2.1.0
types-python-dateutil
typing-utils
typing_extensions
tzdata
ujson==5.9.0
unibox==0.3.20
unicodedata2
uri-template
urllib3
wcwidth
webcolors
webencodings
websocket-client
Werkzeug==3.0.1
widgetsnbextension
xformers==0.0.22.post4+cu118
xyzservices
zipp